change measurement types to floats, add pydrive to ndvi-cam to upload live feed images

+++ api/common/measurements.py
@@ -1,8 +1,8 @@
 ACCEPTED_MEASUREMENTS = [
-  [ "temperature", int ],
-  [ "moisture", int ],
-  [ "humidity", int ],
-  [ "light", int ],
-  [ "water_level", int ],
+  [ "temperature", float ],
+  [ "moisture", float ],
+  [ "humidity", float ],
+  [ "light", float ],
+  [ "water_level", float ],
   [ "light_on", bool ]
 ]
